fix allocHookResrouces signatures
